# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Danshin Genealogy",
        "description": "Доступ ко всем данным только через API токен,\n        в противном случае будет возвращаться ошибка 401.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tree/model/{id}": {
            "get": {
                "operationId": "a9e497e90504ecfe429c4e6b0fc10e97",
                "parameters": [
                    {
                        "name": "id",
                        "in": null,
                        "description": "Id запрашиваемого лица",
                        "required": true
                    },
                    {
                        "name": "parent_id",
                        "in": null,
                        "description": "Id родителя (запрашиваемого лица) по которому строится древо,\n                    по умолчанию древо строится случайным образом по отцу или матери.",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Модель древа запрашиваемого лица в формате JSON."
                    }
                }
            }
        },
        "/api/tree/image/{id}": {
            "get": {
                "operationId": "178879ccc218f6c988c26e0b3e120227",
                "parameters": [
                    {
                        "name": "id",
                        "in": null,
                        "description": "Id запрашиваемого лица",
                        "required": true
                    },
                    {
                        "name": "parent_id",
                        "in": null,
                        "description": "Id родителя (запрашиваемого лица) по которому строится древо,\n                    по умолчанию древо строится случайным образом по отцу или матери.",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Файл изображения, содержащий древо запрашиваемого лица."
                    }
                }
            }
        },
        "/api/tree/image-interactive/{id}": {
            "get": {
                "operationId": "8cf161956d59d8917405007bbcaf2900",
                "parameters": [
                    {
                        "name": "id",
                        "in": null,
                        "description": "Id запрашиваемого лица",
                        "required": true
                    },
                    {
                        "name": "parent_id",
                        "in": null,
                        "description": "Id родителя (запрашиваемого лица) по которому строится древо,\n                    по умолчанию древо строится случайным образом по отцу или матери.",
                        "required": false
                    },
                    {
                        "name": "path_person",
                        "in": null,
                        "description": "Путь к лицу.",
                        "required": true
                    },
                    {
                        "name": "path_tree",
                        "in": null,
                        "description": "Путь к древу лица.",
                        "required": true
                    },
                    {
                        "name": "image_person",
                        "in": null,
                        "description": "Путь к изображению, используемому в ссылке.",
                        "required": true
                    },
                    {
                        "name": "image_tree",
                        "in": null,
                        "description": "Путь к изображению, используемому в ссылке.",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Файл изображения, содержащий древо запрашиваемого лица,\n            содержащий интерактивные ссылки в древе.\n            Ссылка на лицо должна соответствовать шаблону path/person_id,\n            где path - путь к древу,\n            person_id - id номер запрашиваемого лица,\n            то есть Вы предоставляете path, к которому настоящее API добавляет id лица.\n            Ссылка на древо лица должна соответствовать шаблону path/person_id.\n            Также Вы должны предоставить путь к двум изображениям, которые будут использоваться в ссылке."
                    }
                }
            }
        },
        "/api/tree/toggle/{id}": {
            "get": {
                "operationId": "ef419dfded6c6beba9db66a8b4daa6e3",
                "parameters": [
                    {
                        "name": "id",
                        "in": null,
                        "description": "Id запрашиваемого лица",
                        "required": true
                    },
                    {
                        "name": "parent_id",
                        "in": null,
                        "description": "Id родителя (запрашиваемого лица) по которому строится древо,\n                    по умолчанию древо строится случайным образом по отцу или матери.\n                    Id родителя должно иметь тип integer.",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Переключатель для древа запрашиваемого лица."
                    },
                    "204": {
                        "description": "Переключатель для запрашиваемого лица отсутствует,\n                    так как лицо не имеет родителей."
                    },
                    "404": {
                        "description": "Лицо с запрошенным id не существует\n                    или запрошенное лицо не имеет родителя, с переданным parent_id."
                    },
                    "422": {
                        "description": "Параметры запроса не соответствуют ограничениям."
                    }
                }
            }
        }
    }
}